整理了一下命名格式
1.尽量使用驼峰
2.在无用页面前加上$$
