--- README.md
@@ -1,6 +1,0 @@
-# Web
-
-站长：卢瑞元
-技术支持： Paladnix
-
-# 最新的开发进度待站长更新
