--- contactuscontent.php
@@ -1,32 +1,0 @@
-
-	<div id="fh5co-testimonial">
-		<div class="container">
-			<div class="row">
-			建设中
-			<!--
-				<div class="col-md-6 text-center item-block animate-box" data-animate-effect="fadeIn">
-					<blockquote>
-						<p>&ldquo; What we are &rdquo;</p>
-						<p><span class="fh5co-author"><cite>内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容</cite></span></p>
-												<figure>
-							<img src="images/testimonial_person2.jpg" alt="Free Website Template by FreeHTML5.co">
-						</figure>
-						 <p><a href="#" class="btn btn-primary">Learn more</a></p>
-
-					</blockquote>
-				</div>
-				<div class="col-md-6 text-center item-block animate-box" data-animate-effect="fadeIn">
-					<blockquote>
-						<p>&ldquo; What we do &rdquo;</p>
-						<p><span class="fh5co-author"><cite>内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容内容</cite></span></p>
-												<figure>
-							<img src="images/testimonial_person2.jpg" alt="Free Website Template by FreeHTML5.co">
-						</figure>
-						 <p><a href="activity.php" class="btn btn-primary">Learn more</a></p>
-
-					</blockquote>
-				</div>
-				-->
-			</div>
-		</div>
-	</div>
