+++ f4fRepairmanIntro.php
@@ -1,0 +1,55 @@
+<div id="fh5co-agents">
+		<div class="container">
+			<div class="row">
+				<div class="col-md-6 col-md-offset-3 text-center fh5co-heading white animate-box" data-animate-effect="fadeIn">
+					<h2>在线联络员</h2>
+					<p>挑选一个看上去靠谱的小伙子以获得帮助</p>
+				</div>
+				<div class='col-md-12' >
+					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
+						<div class="fh5co-agent">
+							<figure>
+								<img src="images/CproHead1.jpg">
+							</figure>
+							<h3>孙昌来</h3>
+							<p>羞涩</p>
+						<p><a href="f4fRepairmanConnect.php?manid=1&report_id=<?php echo $_GET['report_id']; ?>" class="btn btn-primary btn-outline">Contact Me</a></p> 
+						
+						</div>
+					</div>
+					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
+						<div class="fh5co-agent">
+							<figure>
+								<img src="images/CproHead2.jpg">
+							</figure>
+							<h3>杨增</h3>
+							<p>而不</p>
+							<p><a href="f4fRepairmanConnect.php?manid=2&report_id= <?php echo $_GET['report_id']; ?>" class="btn btn-primary btn-outline">Contact Me</a></p> 
+						</div>
+					</div>
+					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
+						<div class="fh5co-agent">
+							<figure>
+								<img src="images/CproHead3.jpg" >
+							</figure>
+							<h3>王瑞津</h3>
+							<p>想要</p>
+							<p><a href="f4fRepairmanConnect.php?manid=3&report_id= <?php echo $_GET['report_id']; ?>" class="btn btn-primary btn-outline">Contact Me</a></p>
+						</div>
+					</div>
+					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
+						<div class="fh5co-agent">
+							<figure>
+								<img src="images/CproHead4.jpg" >
+							</figure>
+							<h3>党宏波</h3>
+							<p>说话</p>
+						<p><a href="f4fRepairmanConnect.php?manid=4&report_id= <?php echo $_GET['report_id']; ?>" class="btn btn-primary btn-outline">Contact Me</a></p> 
+
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
+</div>
+
