+++ fixforfreeQuestion.php
@@ -1,0 +1,31 @@
+
+	<div id="fh5co-testimonial">
+		<div class="container">
+			<div class="row">
+				<div class="col-md-6 col-md-offset-3 text-center fh5co-heading animate-box" data-animate-effect="fadeIn">
+					<h2>常见问题</h2>
+					<p>在报修之前，我们建议您先按照下面的提示进行自检. </p>
+				</div>
+				<div class="col-md-4 text-center item-block animate-box" data-animate-effect="fadeIn">
+					<blockquote>
+						<p>&ldquo; Step I &rdquo;</p>
+						<p><span class="fh5co-author"><cite>用准确而精炼的词汇确定您的问题</cite></span></p>
+
+					</blockquote>
+				</div>
+				<div class="col-md-4 text-center item-block animate-box" data-animate-effect="fadeIn">
+					<blockquote>
+						<p>&ldquo; Step II &rdquo;</p>
+						<p><span class="fh5co-author"><cite>打开百度</cite></span></p>
+					</blockquote>
+				</div>
+				<div class="col-md-4 text-center item-block animate-box" data-animate-effect="fadeIn">
+						
+					<blockquote>
+						<p>&ldquo; Step III &rdquo;</p>
+						<p><span class="fh5co-author"><cite>使用百度搜索您的问题。</cite></span></p>
+					</blockquote>
+				</div>
+			</div>
+		</div>
+	</div>
