+++ head.php
@@ -9,7 +9,7 @@
 							<li><a href="index.php">首页</a></li>
 							<li><a href="activity.php">活动</a></li>
 							<li><a href="fixforfree.php">电脑义修</a></li>
-							<li class="cta"><a href="contactus.php">关于我们</a></li>
+							<li class="cta"><a href="aboutus.php">关于我们</a></li>
 						</ul>
 					</nav>
 				</div>
