--- repairman_introduc_raw.php
@@ -1,59 +1,0 @@
-<div id="fh5co-agents">
-		<div class="container">
-			<div class="row">
-				<div class="col-md-6 col-md-offset-3 text-center fh5co-heading white animate-box" data-animate-effect="fadeIn">
-					<h2>在线联络员</h2>
-					<p>挑选一个看上去靠谱的小伙子以获得帮助</p>
-				</div>
-				<div >
-					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
-						<div class="fh5co-agent">
-							<figure>
-								<img src="images/CproHead1.jpg">
-							</figure>
-							<h3>孙昌来</h3>
-							<p>羞涩</p>
-						<a target="_blank" href="tencent://message/?uin=774500050&site=127.0.0.1&menu=yes"><img border="0" src="http://wpa.qq.com/pa?p=2:774500050:52" alt="点击这里给我发消息" title="点击这里给我发消息"/></a>
-						
-					<!--<p><a href="#" class="btn btn-primary btn-outline">Contact Me</a></p>-->
-						</div>
-					</div>
-					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
-						<div class="fh5co-agent">
-							<figure>
-								<img src="images/CproHead2.jpg">
-							</figure>
-							<h3>杨增</h3>
-							<p>而不</p>
-						<!--	<p><a href="#" class="btn btn-primary btn-outline">Contact Me</a></p> -->
-					<!--	<a target="_blank" href="http://wpa.qq.com/msgrd?v=3&uin=2359981832&site=qq&menu=yes"><img border="0" src="http://wpa.qq.com/pa?p=2:2359981832:41" alt="点击这里给我发消息" title="点击这里给我发消息"/></a> -->
-					<a target="_blank" href="http://wpa.qq.com/msgrd?v=3&uin=1055614742&site=qq&menu=yes"><img border="0" src="http://wpa.qq.com/pa?p=2:1055614742:52" alt="点击这里给我发消息" title="点击这里给我发消息"/></a>
-						</div>
-					</div>
-					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
-						<div class="fh5co-agent">
-							<figure>
-								<img src="images/CproHead3.jpg" >
-							</figure>
-							<h3>王瑞津</h3>
-							<p>想要</p>
-							<!--<p><a href="#" class="btn btn-primary btn-outline">Contact Me</a></p>-->
-							<a target="_blank" href="http://wpa.qq.com/msgrd?v=3&uin=1361421504&site=qq&menu=yes"><img border="0" src="http://wpa.qq.com/pa?p=2:1361421504:52" alt="点击这里给我发消息" title="点击这里给我发消息"/></a>
-						</div>
-					</div>
-					<div class="col-md-3 text-center item-block animate-box" data-animate-effect="fadeIn">
-						<div class="fh5co-agent">
-							<figure>
-								<img src="images/CproHead4.jpg" >
-							</figure>
-							<h3>党宏波</h3>
-							<p>说话</p>
-						<!--	<p><a href="#" class="btn btn-primary btn-outline">Contact Me</a></p> -->
-						<a target="_blank" href=tencent://message/?uin=627971428&site=127.0.0.1&menu=yes"><img border="0" src="http://wpa.qq.com/pa?p=2:627971428:52" alt="我是早睡早起的好孩纸" title="我是早睡早起的好孩纸"/></a>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-</div>
-
